move some commands
