+++ runway/commands/modules/__init__.py
@@ -2,12 +2,9 @@
 
 # pylint: disable = wildcard-import
 
-from .deploy import * # noqa
-from .destroy import * # noqa
-from .dismantle import * # noqa
-from .plan import *  # noqa
-from .preflight import *  # noqa
-from .takeoff import *  # noqa
-from .taxi import *  # noqa
-from .gitclean import * # noqa
-from .whichenv import *  # noqa
+from .deploy    import *  # noqa
+from .destroy   import *  # noqa
+from .dismantle import *  # noqa
+from .plan      import *  # noqa
+from .takeoff   import *  # noqa
+from .taxi      import *  # noqa
