--- runway/commands/modules/gitclean.py
@@ -1,11 +1,0 @@
-"""The gitclean command."""
-
-from ..modules_command import ModulesCommand
-
-
-class GitClean(ModulesCommand):
-    """Extend ModulesCommand with execute to run the gitclean method."""
-
-    def execute(self):
-        """Clean directory of files."""
-        self.gitclean()
