+++ runway/commands/modules/takeoff.py
@@ -1,10 +1,6 @@
 """The takeoff command."""
-from ..modules_command import ModulesCommand
 
+from .deploy import Deploy
 
-class Takeoff(ModulesCommand):
-    """Extend ModulesCommand with execute to run the deploy method."""
-
-    def execute(self):
-        """Run deployments."""
-        self.deploy()
+# just an alias
+Takeoff = Deploy
