+++ runway/commands/modules/taxi.py
@@ -1,10 +1,6 @@
 """The taxi command."""
-from ..modules_command import ModulesCommand
 
+from .plan import Plan
 
-class Taxi(ModulesCommand):
-    """Extend ModulesCommand with execute to run the plan method."""
-
-    def execute(self):
-        """Generate plans."""
-        self.plan()
+# just an alias
+Taxi = Plan
