+++ runway/commands/modules_command.py
@@ -5,14 +5,11 @@ from __future__ import print_function
 # https://github.com/PyCQA/pylint/issues/73
 from distutils.util import strtobool  # noqa pylint: disable=no-name-in-module,import-error
 
-from subprocess import check_call, check_output
-
 import copy
 import glob
 import json
 import logging
 import os
-import shutil
 import sys
 
 from builtins import input
@@ -21,7 +18,7 @@ import boto3
 import six
 import yaml
 
-from .runway_command import RunwayCommand
+from .runway_command import RunwayCommand, get_env, get_deployment_env_vars
 from ..context import Context
 from ..util import change_dir, load_object_from_string, merge_dicts
 
@@ -80,71 +77,6 @@ def determine_module_class(path, module_options):
                      os.path.basename(path))
         sys.exit(1)
     return load_object_from_string(module_options['class_path'])
-
-
-def get_deployment_env_vars(env_name, env_var_config=None, env_root=None):
-    """Return applicable environment variables."""
-    if env_var_config is None:
-        env_var_config = {}
-    if env_var_config.get('*'):
-        env_vars = env_var_config['*'].copy()
-    else:
-        env_vars = {}
-    if env_var_config.get(env_name):
-        for (key, val) in env_var_config[env_name].items():
-            # Lists are presumed to be path components and will be turned back
-            # to strings
-            if isinstance(val, list):
-                env_var_config[env_name][key] = os.path.join(env_root, os.path.join(*val)) if (env_root and not os.path.isabs(os.path.join(*val))) else os.path.join(*val)  # noqa pylint: disable=line-too-long
-        env_vars = merge_dicts(env_vars, env_var_config[env_name])
-    return env_vars
-
-
-def get_env_from_branch(branch_name):
-    """Determine environment name from git branch name."""
-    if branch_name.startswith('ENV-'):
-        return branch_name[4:]
-    if branch_name == 'master':
-        LOGGER.info('Translating git branch "master" to environment '
-                    '"common"')
-        return 'common'
-    return branch_name
-
-
-def get_env_from_directory(directory_name):
-    """Determine environment name from directory name."""
-    if directory_name.startswith('ENV-'):
-        return directory_name[4:]
-    return directory_name
-
-
-def get_env(path, ignore_git_branch=False):
-    """Determine environment name."""
-    if 'DEPLOY_ENVIRONMENT' in os.environ:
-        return os.environ['DEPLOY_ENVIRONMENT']
-
-    if ignore_git_branch:
-        LOGGER.info('Skipping environment lookup from current git branch '
-                    '("ignore_git_branch" is set to true in the runway '
-                    'config)')
-    else:
-        # These are not located with the top imports because they throw an
-        # error if git isn't installed
-        from git import Repo as GitRepo
-        from git.exc import InvalidGitRepositoryError
-
-        try:
-            b_name = GitRepo(
-                path,
-                search_parent_directories=True
-            ).active_branch.name
-            LOGGER.info('Deriving environment name from git branch %s...',
-                        b_name)
-            return get_env_from_branch(b_name)
-        except InvalidGitRepositoryError:
-            pass
-    LOGGER.info('Deriving environment name from directory %s...', path)
-    return get_env_from_directory(os.path.basename(path))
 
 
 def path_is_current_dir(path):
@@ -304,24 +236,6 @@ def echo_detected_environment(env_name, env_vars):
 class ModulesCommand(RunwayCommand):
     """Env deployment class."""
 
-    def gitclean(self):
-        """Execute git clean to remove untracked/build files."""
-        clean_cmd = ['git', 'clean', '-X', '-d']
-        if 'CI' not in os.environ:
-            print('The following files/directories will be deleted:')
-            print('')
-            print(check_output(clean_cmd + ['-n']).decode())
-            if not strtobool(input('Proceed?: ')):
-                return False
-        check_call(clean_cmd + ['-f'])
-        empty_dirs = self.get_empty_dirs(self.env_root)
-        if empty_dirs != []:
-            print('Now removing empty directories:')
-        for directory in empty_dirs:
-            print("Removing %s/" % directory)
-            shutil.rmtree(os.path.join(self.env_root, directory))
-        return True
-
     def run(self, deployments=None, command='plan'):  # noqa pylint: disable=too-many-branches,too-many-statements
         """Execute apps/code command."""
         if deployments is None:
@@ -340,6 +254,7 @@ class ModulesCommand(RunwayCommand):
         if command == 'destroy':
             LOGGER.info('WARNING!')
             LOGGER.info('Runway is running in DESTROY mode.')
+
         if context.env_vars.get('CI', None):
             if command == 'destroy':
                 deployments_to_run = self.reverse_deployments(deployments)
@@ -401,36 +316,7 @@ class ModulesCommand(RunwayCommand):
                     if deployment.get('current_dir'):
                         modules.append('.' + os.sep)
                     for module in modules:
-                        module_opts = {}
-                        if deployment.get('environments'):
-                            module_opts['environments'] = deployment['environments'].copy()  # noqa
-                        if deployment.get('module_options'):
-                            module_opts['options'] = deployment['module_options'].copy()  # noqa
-                        if isinstance(module, six.string_types):
-                            module = {'path': module}
-                        if path_is_current_dir(module['path']):
-                            module_root = self.env_root
-                        else:
-                            module_root = os.path.join(self.env_root, module['path'])
-                        module_opts = merge_dicts(module_opts, module)
-                        module_opts = load_module_opts_from_file(module_root, module_opts)
-                        if deployment.get('skip-npm-ci'):
-                            module_opts['skip_npm_ci'] = True
-
-                        LOGGER.info("")
-                        LOGGER.info("---- Processing module '%s' for '%s' in %s --------------",
-                                    module['path'],
-                                    context.env_name,
-                                    region)
-                        LOGGER.info("Module options: %s", module_opts)
-                        with change_dir(module_root):
-                            getattr(
-                                determine_module_class(module_root, module_opts)(  # noqa
-                                    context=context,
-                                    path=module_root,
-                                    options=module_opts
-                                ),
-                                command)()
+                        self._deploy_module(module, deployment, context, command)
 
                 if deployment.get('assume-role'):
                     post_deploy_assume_role(deployment['assume-role'], context)
@@ -438,17 +324,45 @@ class ModulesCommand(RunwayCommand):
                 LOGGER.error('No region configured for any deployment')
                 sys.exit(1)
 
-    def plan(self, deployments=None):
-        """Plan apps/code deployment."""
-        self.run(deployments=deployments, command='plan')
+    def _deploy_module(self, module, deployment, context, command):
+        module_opts = {}
+        if deployment.get('environments'):
+            module_opts['environments'] = deployment['environments'].copy()  # noqa
+        if deployment.get('module_options'):
+            module_opts['options'] = deployment['module_options'].copy()  # noqa
+        if isinstance(module, six.string_types):
+            module = {'path': module}
+        if path_is_current_dir(module['path']):
+            module_root = self.env_root
+        else:
+            module_root = os.path.join(self.env_root, module['path'])
+        module_opts = merge_dicts(module_opts, module)
+        module_opts = load_module_opts_from_file(module_root, module_opts)
+        if deployment.get('skip-npm-ci'):
+            module_opts['skip_npm_ci'] = True
 
-    def deploy(self, deployments=None):
-        """Deploy apps/code."""
-        self.run(deployments=deployments, command='deploy')
-
-    def destroy(self, deployments=None):
-        """Deploy apps/code."""
-        self.run(deployments=deployments, command='destroy')
+        LOGGER.info("")
+        LOGGER.info("---- Processing module '%s' for '%s' in %s --------------",
+                    module['path'],
+                    context.env_name,
+                    context.env_region)
+        LOGGER.info("Module options: %s", module_opts)
+        with change_dir(module_root):
+            # dynamically load the particular module's class, 'get' the method
+            #  associated with the command, and call the method
+            module_class = determine_module_class(module_root, module_opts)
+            module_instance = module_class(
+                context=context,
+                path=module_root,
+                options=module_opts
+            )
+            if hasattr(module_instance, command):
+                command_method = getattr(module_instance, command)
+                command_method()
+            else:
+                LOGGER.error("'%s' is missing method '%s'",
+                             module_instance, command)
+                sys.exit(1)
 
     @staticmethod
     def reverse_deployments(deployments=None):
