+++ runway/commands/runway/__init__.py
@@ -2,5 +2,9 @@
 
 # pylint: disable = wildcard-import
 
-from .init import *             # noqa
-from .gen_sample import *       # noqa
+from .gen_sample    import *  # noqa
+from .gitclean      import *  # noqa
+from .init          import *  # noqa
+from .preflight     import *  # noqa
+from .test          import *  # noqa
+from .whichenv      import *  # noqa
