+++ runway/commands/runway/preflight.py
@@ -1,0 +1,5 @@
+"""The preflight command."""
+from .test import Test
+
+# just an classname alias
+Preflight = Test
